% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_reg.R
\name{compute_integral}
\alias{compute_integral}
\title{monte carlo integral of non_param_cop (using importance sampling)}
\usage{
compute_integral(n_samples, user_info = FALSE)
}
\arguments{
\item{n_samples:}{Number of samples to create to compute the integral}

\item{user_info:}{Whether to display an information, which steps are currently running.}
}
\value{
The monte carlo integral approximation.
}
\description{
monte carlo integral of non_param_cop (using importance sampling)
}
