% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_reg.R
\name{non_param_cop}
\alias{non_param_cop}
\title{compute the value of the non-parametric copula obtained from the simplified fit,
together with the classifier}
\usage{
non_param_cop(obs, nu = 1)
}
\arguments{
\item{obs}{the observations, for which the non-parametric copula should be created.}

\item{nu}{The fraction of noise to real samples}
}
\value{
The non-parametric copula evaluated at the points obs.
}
\description{
compute the value of the non-parametric copula obtained from the simplified fit,
together with the classifier
}
