% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_non_simplified_vine.R
\name{pairs_copula_data_custom}
\alias{pairs_copula_data_custom}
\title{Function with similar functionality as the pairs_copula_data function of
the rvinecopulib package. Plots histograms of df on the diagonal,
scatter plots on the upper triangle and
marginally normalized contours on the lower triangle.}
\usage{
pairs_copula_data_custom(u_data, max_samples = -1, plot_emp_ktau = FALSE)
}
\arguments{
\item{u_data}{Data to plot, each dimension needs to be between 0 and 1}

\item{max_samples}{Maximum number of samples to consider for the plot
If -1, then all samples are considered.
If >0, then a random sample of max_samples is drawn and considered for the plots}

\item{plot_emp_ktau}{Whether to plot the empirical kendall's tau (TRUE or FALSE)}
}
\description{
Function with similar functionality as the pairs_copula_data function of
the rvinecopulib package. Plots histograms of df on the diagonal,
scatter plots on the upper triangle and
marginally normalized contours on the lower triangle.
}
