% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_methods.R
\name{perform_quant_reg}
\alias{perform_quant_reg}
\title{performs a quantile regression}
\usage{
perform_quant_reg(
  g_vals,
  orig_data,
  family_set_name = "onepar",
  bottom_quantile_levels = c(0.05, 0.1),
  top_quantile_levels = c(0.9, 0.95)
)
}
\arguments{
\item{g_vals:}{Vector. Values g_t from the thesis}

\item{orig_data:}{The observed data.}

\item{family_set_name:}{Defaults to "onepar".
Argument family_set in the D-vine Quantile regression function.
Possible choices include "onepar" and "parametric".}

\item{bottom_quantile_levels:}{Defaults to c(0.05,0.1). Vector of lower quantiles,
for which it is checked, whether the conditioned quantile estimates are >0.
Tests whether the alternative model is better}

\item{top_quantile_levels:}{Defaults to c(0.05,0.1). Vector of upper quantiles,
for which it is checked, whether the conditioned quantile estimates are <0.
Tests whether the simplified model is better}
}
\value{
: A List of 2 vectors, the first containing the number of samples, for which
q(bottom_quantile_levels) > 0 holds, the second containing the number of samples, for
which q(top_quantile_levels) < 0 holds.
}
\description{
performs a quantile regression
}
