% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_non_simplified_vine.R
\name{permute_indices}
\alias{permute_indices}
\title{permutes indices of a rvine structure matrix so that on the antidiagonal
the values 1 to d (dimension of the matrix) are ordered from top right to bottom left}
\usage{
permute_indices(mat)
}
\arguments{
\item{mat:}{The rvine structure matrix}
}
\description{
permutes indices of a rvine structure matrix so that on the antidiagonal
the values 1 to d (dimension of the matrix) are ordered from top right to bottom left
}
