% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_experiments.R
\name{run_simulations}
\alias{run_simulations}
\title{Run the simulations}
\usage{
run_simulations(
  num_samples,
  dims,
  nus,
  tau_limits,
  struct_mats,
  families,
  initial_params,
  rotations,
  filename = ""
)
}
\arguments{
\item{num_samples:}{list of integers, that determine for what different
number of samples drawn from a non-simplified vine copula the experiments should be run.}

\item{nus:}{list of numbers: Determines how many noise (simplified) samples should be simulated.
The i-th entry contains the values nu = n_{noise}/n_{true} for
the number of samples specified in num_samples[[i]]}

\item{dims:}{list of integers, that determine for what dimensions to run the tests
(currently implemented for 3, 4 and 5)}

\item{tau_limits:}{list of 2 dimensional vectors: For each of those vectors,
the first entry determines, what the lower threshold of kendalls tau is,
the second entry determines what the upper threshold of kendalls tau is in the
conditional copulas}

\item{struct_mats:}{List of regular vine matrices: Determine what structure to use.
The entry in position i needs to have the same dimension as dims[[i]].}

\item{families:}{List of (list of list of string): The i-th element
contains the copula families corresponding to the copulas defined with struct_mats[[i]]}

\item{initial_params:}{List of vectors: The i-th entry contains the parameters
for the i-th unconditioned copula in the first tree, defined by struct_mats[[i]]}

\item{rotations:}{List of (list of list of int): The i-th element contains the
rotations of the copulas specified in struct_mats[[i]]}

\item{filename:}{string, defaults to "": If not "", then the results are written to
a csv file, with name specified as filename}
}
\value{
result_df: A Dataframe with the results of the experiments.
}
\description{
Run the simulations
}
