% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_non_simplified_vine.R
\name{simulate_3d}
\alias{simulate_3d}
\title{Simulates a 3d non-simplified r vine copula}
\usage{
simulate_3d(
  n_samples = 500,
  struct = matrix(c(1, 1, 1, 2, 2, 0, 3, 0, 0), byrow = TRUE, ncol = 3),
  families = list(list("gumbel", "gumbel"), list("gumbel")),
  params = list(c(2), c(1.3)),
  param_cond_func = u_to_param,
  rotations = list(list(0, 0), list(0))
)
}
\arguments{
\item{n_samples:}{Number of Samples}

\item{struct:}{r-vine structure matrix}

\item{families:}{families of the copulas to use.
Formatted as a list of lists, e.g.list[[1]] gives the list containing the pair copulas in the first tree
(From the first column on the left to the second-last column of struct).}

\item{params:}{list containing the parameters of the families for the first tree}

\item{param_cond_func:}{A function computing the parameter of the copula in the second tree, from
the sampled value of the conditioning variable}

\item{rotations:}{Rotations of the copulas to use. Formatted as a list of lists, similar to the families.}
}
\description{
Simulates a 3d non-simplified r vine copula
}
