% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_non_simplified_vine.R
\name{simulate_single_sample}
\alias{simulate_single_sample}
\title{Simulates a single sample,
This function is called by simulate_non_simplified, after modifying the input structure,
to be able to parallelize the procedure more easily for better performance}
\usage{
simulate_single_sample(
  new_struct,
  M_mat,
  families,
  params,
  param_cond_func,
  rotations
)
}
\arguments{
\item{new_struct}{rvine structure, where the antidiagonal is ordered
increasingly from top right to bottom left.}

\item{M_mat}{the matrix \tilde{M}, with \tilde{M_{k,i}} = max_{j=1}^{k}{new_struct_{j,i}}}

\item{families}{copula families, formatted as a list of lists}

\item{params}{parameters for the first tree, formatted as a list (of vectors)}

\item{param_cond_func}{function that maps the u values, which the respective
copula is conditioned on, to a parameter for that copula (needs to be able to handle
different lengths of input from 1 to d-2)}

\item{rotations}{rotations of the copulas, formatted as a list of lists.}
}
\description{
Simulates a single sample,
This function is called by simulate_non_simplified, after modifying the input structure,
to be able to parallelize the procedure more easily for better performance
}
