% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_methods.R
\name{train_model}
\alias{train_model}
\title{function to train the model}
\usage{
train_model(
  model,
  x_train,
  y_train,
  lr_schedule = lr_schedule_fun,
  num_epochs = 300,
  batch_size = 128,
  val_split = 0.2,
  verbose = 1
)
}
\arguments{
\item{model:}{The NN model to train}

\item{x_train:}{training data}

\item{y_train:}{labels for the training data}

\item{lr_schedule:}{Function for learning rate scheduling. Defaults to lr_schedule_fun.}

\item{num_epochs:}{Integer, defaults to 500.
Number of epochs for which the model should be trained.}

\item{batch_size:}{Integer, defaults to 128.
Batch size to use to compute gradients during training}

\item{val_split:}{Number between 0 and 1. Fraction of data used for validation.}

\item{verbose:}{0 or 1, defaults to 1.
If 1, an output is printed after every epoch of the training process.
If 0, no output is printed.}
}
\value{
model: the trained model
}
\description{
function to train the model
}
