% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_methods.R
\name{train_test_split}
\alias{train_test_split}
\title{Performs a train test split
Gives all datapoints in orig_data a label of 1, all datapoints in simplified data a label of 0}
\usage{
train_test_split(
  orig_data,
  simplified_data,
  train_perc = 0.8,
  stratified = FALSE
)
}
\arguments{
\item{orig_data:}{The real dataset, which gets labels of 1}

\item{simplified_data:}{Synthetic dataset, simulated from a simplified
vine copula, gets labels of 0}

\item{train_perc:}{Number between 0 and 1, defaults to 0.8.
Determins what fraction of the data is used for training.}

\item{stratified:}{TRUE or FALSE, defaults to FALSE
Whether the share of points from orig data should be exactly the same in train and test set.}
}
\value{
A list with 4 entries, x_train, x_test, y_train, y_test in that order.
}
\description{
Performs a train test split
Gives all datapoints in orig_data a label of 1, all datapoints in simplified data a label of 0
}
