% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_non_simplified_vine.R
\name{u_to_param}
\alias{u_to_param}
\title{maps a u value to the specific parameter for the third copula, by
finding a value on the fisher-z-transform-scale and mapping that back to a parameter}
\usage{
u_to_param(u, family = "gaussian")
}
\arguments{
\item{u:}{the samples u value (between 0 and 1)}

\item{family:}{the family of the third copula in the sample
#' @returns A function of u and family, which calculates the parameter of
a copula given the conditioned values.}
}
\description{
maps a u value to the specific parameter for the third copula, by
finding a value on the fisher-z-transform-scale and mapping that back to a parameter
}
