% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_non_simplified_vine.R
\name{u_to_param_non_lin}
\alias{u_to_param_non_lin}
\title{takes a vector a and returns a function of u (vector of elements between 0 and 1)
and of a string "family", which corresponds to a copula family.
that function transforms the values u using qnorm (inverse normal dist.function)
and then builds polynomial terms of degree 2, if dim(u) <=3, else a linear function of these
with a as parameters.}
\usage{
u_to_param_non_lin(a, tau_lower = -0.92, tau_upper = 0.92)
}
\arguments{
\item{a:}{a vector of weights}

\item{tau_lower:}{Lowest tau value
(should be between -1 and 1 and less than tau_upper)}

\item{tau_upper:}{Highest tau value
(should be between -1 and 1 and greater than tau_lower)}
}
\value{
A function of u and family, which calculates the parameter of
a copula given the conditioned values.
}
\description{
takes a vector a and returns a function of u (vector of elements between 0 and 1)
and of a string "family", which corresponds to a copula family.
that function transforms the values u using qnorm (inverse normal dist.function)
and then builds polynomial terms of degree 2, if dim(u) <=3, else a linear function of these
with a as parameters.
}
